Co-locate decomposition CLI with library module
